polish: add dp kind specification to real declarations (fixes #2581)
Add explicit real(dp) type specification to 18 real declarations
across memory, performance, semantic, and AST subsystems per
CLAUDE.md Fortran rules:

- src/memory/arena_memory.f90: 1 type field + 1 literal
- src/memory/compiler_arena.f90: 3 type fields + 7 literals
- src/performance/ast_performance.f90: 5 type fields + 2 literals
- src/semantic/types/type_system_arena.f90: 1 type field
- src/semantic/convergence_checker.f90: 2 type fields
- src/semantic/analyzers/base_analyzer.f90: 1 type field
- src/ast/arena/ast_arena_core.f90: 4 type fields
- src/ast/ast_base.f90: 1 type field
- src/ast/ast_node_registry.f90: 1 type field

This ensures numerical precision consistency by using double
precision (real64) for all floating-point metrics and statistics.

+++ src/semantic/types/type_system_arena.f90
@@ -4,6 +4,7 @@ module type_system_arena
     ! Provides 10-100x performance improvement for type inference operations
 
     use arena_memory
+    use, intrinsic :: iso_fortran_env, only: dp => real64
     implicit none
     private
 
@@ -95,7 +96,7 @@ module type_system_arena
         integer :: poly_types = 0
         integer :: arg_arrays = 0
         integer :: total_memory = 0
-        real :: utilization = 0.0
+        real(dp) :: utilization = 0.0_dp
     end type type_arena_stats_t
 
 contains
